# docs/api/openAPI.yaml
components:
  schemas:
...
    AddressDto:
      additionalProperties: false
      properties:
        bpna:
          maxItems: 50
          pattern: ^BPNA[0-9a-zA-Z]{12}$
          type: string
        country:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        streetAndNumber:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        zipCodeAndCity:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      type: object
...
    Classification:
      additionalProperties: false
      properties:
        classificationDescription:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationID:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationStandard:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      required:
      - classificationID
      - classificationStandard
      type: object
    DataExchangeApprovalDto:
      additionalProperties: false
      properties:
        approvalId:
          maxItems: 50
          type: string
        approvedTypes:
          items:
            enum:
            - n-tier
            maxItems: 50
            type: string
          maxItems: 50
          type: array
        dataExchangeRequestId:
          maxItems: 50
          type: string
        finalized:
          type: boolean
        timestamp:
          format: date-time
          maxItems: 50
          type: string
        uuid:
          format: uuid
          maxItems: 50
          type: string
      required:
      - approvedTypes
      - dataExchangeRequestId
      type: object
    DataExchangeRequestDto:
      additionalProperties: false
      properties:
        criticality:
          enum:
          - high
          - medium
          - low
          maxItems: 50
          type: string
        desiredEndDateTime:
          format: date-time
          maxItems: 50
          type: string
        desiredStartDateTime:
          format: date-time
          maxItems: 50
          type: string
        notificationId:
          format: uuid
          maxItems: 50
          type: string
        requestId:
          maxItems: 50
          type: string
        requestedTypes:
          items:
            enum:
            - n-tier
            maxItems: 50
            type: string
          maxItems: 50
          type: array
        text:
          maxItems: 50
          type: string
        timestamp:
          format: date-time
          maxItems: 50
          type: string
        uuid:
          format: uuid
          maxItems: 50
          type: string
      required:
      - criticality
      - desiredEndDateTime
      - desiredStartDateTime
      - notificationId
      - requestedTypes
      - text
      type: object
...
    PartSitesInformationAsPlanned:
      additionalProperties: false
      properties:
        catenaXsiteId:
          maxItems: 50
          pattern: ^BPNS[0-9a-zA-Z]{12}$
          type: string
        function:
          enum:
          - production
          - warehouse
          - spare part warehouse
          maxItems: 50
          type: string
        functionValidFrom:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
        functionValidUntil:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
      required:
      - catenaXsiteId
      - function
      type: object
    PartTypeInformationBody:
      additionalProperties: false
      properties:
        manufacturerPartId:
          maxItems: 50
          type: string
        nameAtManufacturer:
          maxItems: 50
          type: string
        partClassification:
          items:
            $ref: '#/components/schemas/Classification'
          maxItems: 50
          type: array
      required:
      - manufacturerPartId
      - nameAtManufacturer
      type: object
    PartTypeInformationSAMM:
      additionalProperties: false
      properties:
        catenaXId:
          maxItems: 50
          pattern: (^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)|(^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)
          type: string
        partSitesInformationAsPlanned:
          items:
            $ref: '#/components/schemas/PartSitesInformationAsPlanned'
          maxItems: 50
          type: array
          uniqueItems: true
        partTypeInformation:
          $ref: '#/components/schemas/PartTypeInformationBody'
      required:
      - catenaXId
      - partTypeInformation
      type: object
...
    PartnerDto:
      additionalProperties: false
      properties:
        addresses:
          items:
            $ref: '#/components/schemas/AddressDto'
          maxItems: 50
          type: array
          uniqueItems: true
        bpnl:
          maxItems: 50
          pattern: ^BPNL[0-9a-zA-Z]{12}$
          type: string
        edcUrl:
          maxItems: 50
          pattern: ^http[s]?://([a-z0-9][a-z0-9\-]+[a-z0-9])(\.[a-z0-9\-]+)*(:[0-9]{1,4})?(/[a-z0-9\-]+)*[/]?$
          type: string
        name:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        sites:
          items:
            $ref: '#/components/schemas/SiteDto'
          maxItems: 50
          type: array
          uniqueItems: true
        uuid:
          format: uuid
          maxItems: 50
          type: string
      type: object
...
    SiteDto:
      additionalProperties: false
      properties:
        addresses:
          items:
            $ref: '#/components/schemas/AddressDto'
          maxItems: 50
          type: array
          uniqueItems: true
        bpns:
          maxItems: 50
          pattern: ^BPNS[0-9a-zA-Z]{12}$
          type: string
        name:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      type: object
...
  securitySchemes:
    X-API-KEY:
      in: header
      name: X-API-KEY
      type: apiKey
info:
  title: PURIS FOSS Open API
  version: 6.0.0
openapi: 3.1.0
paths:
...
  /data-exchange-request:
    post:
      description: "Creates a new own data exchange request. \n"
      operationId: createDataExchangeRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataExchangeRequestDto'
              additionalProperties: false
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DataExchangeRequestDto'
                additionalProperties: false
          description: Own Data Exchange Request was created.
        '400':
          description: Malformed or invalid request body.
        '409':
          description: Own Data Exchange Request already exists.
        '500':
          description: Internal Server Error.
      summary: Creates a new own data exchange request
      tags:
      - data-exchange-request-controller
...
  /data-exchange-request/{id}/approval:
    post:
      description: "Creates a new own data exchange approval in response to an existing\
        \ ReportedDataExchangeRequest. \n"
      operationId: createDataExchangeApproval
      parameters:
      - in: path
        name: id
        required: true
        schema:
          additionalProperties: false
          format: uuid
          maxItems: 50
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataExchangeApprovalDto'
              additionalProperties: false
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DataExchangeApprovalDto'
                additionalProperties: false
          description: Own Data Exchange Approval was created.
        '400':
          description: Malformed or invalid request body.
        '409':
          description: Own Data Exchange Approval already exists.
        '500':
          description: Internal Server Error.
      summary: Creates a new own data exchange approval
      tags:
      - data-exchange-request-controller
  /data-exchange-request/{id}/approved-reported:
    get:
      description: Get all reported data exchange approvals.
      operationId: getReportedDataExchangeApproval
      parameters:
      - in: path
        name: id
        required: true
        schema:
          additionalProperties: false
          format: uuid
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DataExchangeApprovalDto'
                additionalProperties: false
          description: OK
      summary: Get all reported data exchange approvals
      tags:
      - data-exchange-request-controller
...
  /partners:
    get:
      description: Returns the requested PartnerDto.
      operationId: getPartner
      parameters:
      - description: The unique BPNL that was assigned to that Partner.
        example: BPNL2222222222RR
        in: query
        name: partnerBpnl
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartnerDto'
                additionalProperties: false
          description: Found Partner, returning it in response body.
        '400':
          description: Invalid parameter.
        '404':
          description: Requested Partner not found.
        '500':
          description: Internal Server Error.
      summary: Gets a specific Partner -- ADMIN ONLY
      tags:
      - partner-controller
    post:
      description: Creates a new Partner entity with the data given in the request
        body. Please note that no UUID can be assigned to a Partner that wasn't created
        before. So the request body must not contain a UUID.
      operationId: createPartner
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerDto'
              additionalProperties: false
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Partner created successfully.
        '400':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Request body was malformed, didn't meet the minimum constraints
            or wrongfully contained a UUID.
        '409':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: 'The BPNL specified in the request body is already assigned. '
      summary: Creates a new Partner -- ADMIN ONLY
      tags:
      - partner-controller
...
  /parttypeinformation/{materialnumber}/submodel/{representation}:
    get:
      description: 'Endpoint that delivers PartTypeInformation of own products to
        customer partners. ''materialnumber'' must be set to the ownMaterialNumber
        of the party, that receives the request. Please note that the SAMMs delivered
        by this endpoint don''t provide partClassification and partSitesInformationAsPlanned
        data. This endpoint is meant to be accessed by partners via EDC only. '
      operationId: getMapping
      parameters:
      - in: header
        name: edc-bpn
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: The material number that the request receiving party uses for
          the material in question
        in: path
        name: materialnumber
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: Must be set to '$value'
        in: path
        name: representation
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartTypeInformationSAMM'
                additionalProperties: false
          description: Ok
        '400':
          description: 'Invalid request parameters. '
        '401':
          description: 'Access forbidden. '
        '404':
          description: 'Product not found for given parameters. '
        '501':
          description: 'Unsupported representation requested. '
      tags:
      - part-type-information-controller
...
security:
- X-API-KEY: []
servers:
- description: Generated server url
  url: http://localhost:8081/catena